Add Home Activity and lots of changes

+++ app/src/main/java/com/example/openrestaurant/MainActivity.kt
@@ -21,7 +21,7 @@ import io.paperdb.Paper
 
 class MainActivity : AppCompatActivity() {
 
-    private val TAG="Main_ACTIVITY"
+    private val TAG = "Main_ACTIVITY"
     private lateinit var fusedLocationClient: FusedLocationProviderClient
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -39,22 +39,23 @@ class MainActivity : AppCompatActivity() {
         }
 
         findViewById<Button>(R.id.businessLoginButton).setOnClickListener {
-//            Toast.makeText(this, "Feature Coming Soon!", Toast.LENGTH_SHORT).show()
             startActivity(Intent(this, QRCodeActivity::class.java))
         }
+
         findViewById<Button>(R.id.btnCreateBusiness).setOnClickListener {
             Toast.makeText(this, "Feature Coming Soon!", Toast.LENGTH_SHORT).show()
         }
 
-        if(ActivityCompat.checkSelfPermission(this,Manifest.permission.ACCESS_COARSE_LOCATION)!= PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(this,Manifest.permission.ACCESS_FINE_LOCATION)!= PackageManager.PERMISSION_GRANTED)
-        {
-            ActivityCompat.requestPermissions(this,arrayOf(Manifest.permission.ACCESS_FINE_LOCATION,Manifest.permission.ACCESS_COARSE_LOCATION),54)
+        if (ActivityCompat.checkSelfPermission(this,
+                Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(
+                this,
+                Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED
+        ) {
+            ActivityCompat.requestPermissions(this,
+                arrayOf(Manifest.permission.ACCESS_FINE_LOCATION,
+                    Manifest.permission.ACCESS_COARSE_LOCATION),
+                54)
         }
-
-
-
-
-
 
 
         fusedLocationClient = LocationServices.getFusedLocationProviderClient(this)
@@ -64,8 +65,9 @@ class MainActivity : AppCompatActivity() {
         request.interval = 10000
         request.fastestInterval = 5000
         request.priority = LocationRequest.PRIORITY_HIGH_ACCURACY
-        val permission = ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION
-        )
+        val permission =
+            ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION
+            )
         if (permission == PackageManager.PERMISSION_GRANTED) {
             fusedLocationClient.requestLocationUpdates(request, object : LocationCallback() {
                 @RequiresApi(Build.VERSION_CODES.O)
@@ -73,8 +75,8 @@ class MainActivity : AppCompatActivity() {
                     val location: Location? = locationResult.lastLocation
                     if (location != null) {
 
-                        Paper.book().write("LATITUDE", location.latitude.toDouble());
-                        Paper.book().write("LONGITUDE", location.longitude.toDouble());
+                        Paper.book().write("LATITUDE", location.latitude);
+                        Paper.book().write("LONGITUDE", location.longitude);
 
 
                         Log.d("longitude ", location.longitude.toString())
@@ -92,22 +94,16 @@ class MainActivity : AppCompatActivity() {
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         super.onActivityResult(requestCode, resultCode, data)
-        if(requestCode== Activity.RESULT_OK){
-            if(requestCode==Iconstant.DEFAULTS.GPS_CODE){
-                Log.d(TAG,"OnActivityResult: SUCCESS")
-            }
-            else{
+        if (requestCode == Activity.RESULT_OK) {
+            if (requestCode == Iconstant.DEFAULTS.GPS_CODE) {
+                Log.d(TAG, "OnActivityResult: SUCCESS")
+            } else {
                 GPSutility(this).turnOnGPS()
             }
-        }
-        else{
+        } else {
             GPSutility(this).turnOnGPS()
         }
     }
-
-
-
-
 
 
 }
